Consolidate /create and /templates
